+++ pydidas/multiprocessing/app_processor_without_tasks_.py
@@ -1,6 +1,6 @@
 # This file is part of pydidas.
 #
-# Copyright 2023, Helmholtz-Zentrum Hereon
+# Copyright 2023 - 2024, Helmholtz-Zentrum Hereon
 # SPDX-License-Identifier: GPL-3.0-only
 #
 # pydidas is free software: you can redistribute it and/or modify
@@ -22,7 +22,7 @@ pre-defined task list.
 """
 
 __author__ = "Malte Storm"
-__copyright__ = "Copyright 2023, Helmholtz-Zentrum Hereon"
+__copyright__ = "Copyright 2023 - 2024, Helmholtz-Zentrum Hereon"
 __license__ = "GPL-3.0-only"
 __maintainer__ = "Malte Storm"
 __status__ = "Production"
@@ -31,7 +31,6 @@ __all__ = ["app_processor_without_tasks"]
 
 import queue
 import time
-from multiprocessing import Queue
 
 from ..core import ParameterCollection
 from ..core.utils import LOGGING_LEVEL, pydidas_logger
@@ -41,10 +40,7 @@ logger = pydidas_logger()
 
 
 def app_processor_without_tasks(
-    input_queue: Queue,
-    output_queue: Queue,
-    stop_queue: Queue,
-    aborted_queue: Queue,
+    multiprocessing_config: dict,
     app: type,
     app_params: ParameterCollection,
     app_config: dict,
@@ -60,16 +56,9 @@ def app_processor_without_tasks(
 
     Parameters
     ----------
-    input_queue : multiprocessing.Queue
-        The input queue which supplies the processor with indices to be
-        processed.
-    output_queue : multiprocessing.Queue
-        The queue for transmissing the results to the controlling thread.
-    stop_queue : multiprocessing.Queue
-        The queue for sending a termination signal to the worker.
-    aborted_queue : multiprocessing.Queue
-        The queue which is used by the processor to signal the calling
-        thread that it has aborted its cycle.
+    multiprocessing_config : dict
+        The multiprocessing configuration dictionary. It includes  information
+        about the logging level as well as queue objects.
     app : BaseApp
         The Application class to be called in the process. The App must have a
         multiprocessing_func method.
@@ -89,7 +78,11 @@ def app_processor_without_tasks(
             level.
     """
     _wait_for_output = kwargs.get("wait_for_output_queue", True)
-    logger.setLevel(kwargs.get("logging_level", LOGGING_LEVEL))
+    logger.setLevel(multiprocessing_config.get("logging_level", LOGGING_LEVEL))
+    _input_queue = multiprocessing_config.get("queue_input")
+    _output_queue = multiprocessing_config.get("queue_output")
+    _stop_queue = multiprocessing_config.get("queue_stop")
+    _aborted_queue = multiprocessing_config.get("queue_aborted")
     _app_carryon = True
     logger.debug("Started process")
     _app = app(app_params, slave_mode=True)
@@ -98,9 +91,9 @@ def app_processor_without_tasks(
     while True:
         # check for stop signal
         try:
-            stop_queue.get_nowait()
+            _stop_queue.get_nowait()
             logger.debug("Received stop queue signal")
-            aborted_queue.put(1)
+            _aborted_queue.put(1)
             _wait_for_output = False
             break
         except queue.Empty:
@@ -110,12 +103,12 @@ def app_processor_without_tasks(
         _app_carryon = _app.multiprocessing_carryon()
         if _app_carryon:
             _index, _results = _app.multiprocessing_func(-1)
-            output_queue.put([_index, _results])
+            _output_queue.put([_index, _results])
         else:
             time.sleep(0.005)
     logger.debug("Worker finished with all tasks.")
     _carry_on = False
-    while _wait_for_output and not output_queue.empty():
+    while _wait_for_output and not _output_queue.empty():
         if not _carry_on:
             logger.debug("Waiting for output queue to empty.")
             _carry_on = True