+++ tests/multiprocessing/test_app_processor_wo_tasks.py
@@ -1,6 +1,6 @@
 # This file is part of pydidas.
 #
-# Copyright 2023, Helmholtz-Zentrum Hereon
+# Copyright 2023 - 2024, Helmholtz-Zentrum Hereon
 # SPDX-License-Identifier: GPL-3.0-only
 #
 # pydidas is free software: you can redistribute it and/or modify
@@ -18,7 +18,7 @@
 """Unit tests for pydidas modules."""
 
 __author__ = "Malte Storm"
-__copyright__ = "Copyright 2023, Helmholtz-Zentrum Hereon"
+__copyright__ = "Copyright 2023 - 2024, Helmholtz-Zentrum Hereon"
 __license__ = "GPL-3.0-only"
 __maintainer__ = "Malte Storm"
 __status__ = "Production"
@@ -41,29 +41,20 @@ class _ProcThread(threading.Thread):
 
     def __init__(
         self,
-        input_queue,
-        output_queue,
-        stop_queue,
-        finished_queue,
+        mp_config: dict,
         app,
         app_params,
         app_config,
     ):
         super().__init__()
-        self.input_queue = input_queue
-        self.output_queue = output_queue
-        self.stop_queue = stop_queue
-        self.finished_queue = finished_queue
+        self._mp_config = mp_config
         self.app = app
         self.app_params = app_params
         self.app_config = app_config
 
     def run(self):
         app_processor_without_tasks(
-            self.input_queue,
-            self.output_queue,
-            self.stop_queue,
-            self.finished_queue,
+            self._mp_config,
             self.app,
             self.app_params,
             self.app_config,
@@ -83,26 +74,25 @@ class Test_app_processor_without_tasks(unittest.TestCase):
         cls._qtapp.quit()
 
     def setUp(self):
-        self.input_queue = mp.Queue()
-        self.output_queue = mp.Queue()
-        self.stop_queue = mp.Queue()
-        self.finished_queue = mp.Queue()
+        self._mp_config = {
+            "queue_input": mp.Queue(),
+            "queue_output": mp.Queue(),
+            "queue_stop": mp.Queue(),
+            "queue_aborted": mp.Queue(),
+        }
         self.app = MpTestAppWoTasks()
         self.n_test = self.app._config["max_index"]
         self.app.multiprocessing_pre_run()
 
     def tearDown(self):
-        self.input_queue.close()
-        self.output_queue.close()
-        self.stop_queue.close()
-        self.finished_queue.close()
+        self._mp_config["queue_input"].close()
+        self._mp_config["queue_output"].close()
+        self._mp_config["queue_stop"].close()
+        self._mp_config["queue_aborted"].close()
 
     def create_thread(self):
         return _ProcThread(
-            self.input_queue,
-            self.output_queue,
-            self.stop_queue,
-            self.finished_queue,
+            self._mp_config,
             self.app.__class__,
             self.app.params.copy(),
             self.app._config,
@@ -113,7 +103,7 @@ class Test_app_processor_without_tasks(unittest.TestCase):
         _tasks = []
         while True:
             try:
-                _index, _item = self.output_queue.get_nowait()
+                _index, _item = self._mp_config["queue_output"].get_nowait()
                 _results.append(_item)
                 _tasks.append(_index)
             except queue.Empty:
@@ -124,21 +114,21 @@ class Test_app_processor_without_tasks(unittest.TestCase):
         _thread = self.create_thread()
         _thread.start()
         time.sleep(0.05)
-        self.stop_queue.put(1)
+        self._mp_config["queue_stop"].put(1)
         time.sleep(0.05)
         _tasks, _results = self.get_results()
         time.sleep(0.01)
         self.assertTrue(len(_tasks) > 0)
-        self.assertEqual(self.finished_queue.get_nowait(), 1)
+        self.assertEqual(self._mp_config["queue_aborted"].get_nowait(), 1)
 
     def test_stop_signal(self):
         _thread = self.create_thread()
         _thread.start()
         time.sleep(0.05)
         self.assertTrue(_thread.is_alive())
-        self.stop_queue.put(1)
+        self._mp_config["queue_stop"].put(1)
         time.sleep(0.05)
-        self.assertEqual(self.finished_queue.get(), 1)
+        self.assertEqual(self._mp_config["queue_aborted"].get(), 1)
 
 
 if __name__ == "__main__":